Backend/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiohttp-retry==2.9.1
aiosignal==1.3.1
alembic==1.14.1
async-timeout==5.0.1
attrs==25.3.0
blinker==1.8.2
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.1.8
cloudinary==1.44.1
ecdsa==0.19.1
Flask==3.0.3
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
frozenlist==1.5.0
greenlet==3.1.1
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==2.1.5
multidict==6.1.0
propcache==0.2.0
PyJWT==2.9.0
python-dotenv==1.0.1
python-http-client==3.3.7
requests==2.32.4
sendgrid==6.12.4
six==1.17.0
SQLAlchemy==2.0.41
twilio==9.6.5
typing_extensions==4.13.2
urllib3==2.2.3
Werkzeug==3.0.6
yarl==1.15.2
zipp==3.20.2
